+++ .idea/dictionaries/project.xml
@@ -1,0 +1,8 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="project">
+    <words>
+      <w>alphanum</w>
+      <w>zabcdefghijklmnopqrstuvwxyz</w>
+    </words>
+  </dictionary>
+</component>
