+++ app/build.gradle
@@ -5,7 +5,8 @@ import com.android.tools.profgen.DexFile
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
-    id("org.jetbrains.kotlin.plugin.compose") version "2.0.21"
+    id("org.jetbrains.kotlin.plugin.compose")
+    id("org.jetbrains.kotlin.plugin.serialization") version "2.3.0"
 }
 
 android {
@@ -23,6 +24,8 @@ android {
         vectorDrawables {
             useSupportLibrary true
         }
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -33,11 +36,8 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_17
-        targetCompatibility JavaVersion.VERSION_17
-    }
-    kotlinOptions {
-        jvmTarget = "17"
+        sourceCompatibility JavaVersion.VERSION_21
+        targetCompatibility JavaVersion.VERSION_21
     }
     buildFeatures {
         compose true
@@ -66,27 +66,30 @@ dependencies {
 //  For Security
     implementation "androidx.biometric:biometric-ktx:1.2.0-alpha05"
     implementation "androidx.security:security-crypto:1.1.0"
+    implementation("androidx.datastore:datastore:1.2.0")
+    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.9.0")
 
 //  For Audio/Video
-    implementation 'androidx.media3:media3-exoplayer:1.8.0'
-    implementation 'androidx.media3:media3-ui:1.8.0'
-    implementation 'androidx.media3:media3-common:1.8.0'
+    implementation 'androidx.media3:media3-exoplayer:1.9.0'
+    implementation 'androidx.media3:media3-ui:1.9.0'
+    implementation 'androidx.media3:media3-common:1.9.0'
 
 //  For Pictures
     implementation 'com.github.bumptech.glide:glide:5.0.5'
     implementation "com.github.bumptech.glide:compose:1.0.0-beta08"
     implementation 'androidx.preference:preference-ktx:1.2.1'
-    implementation 'androidx.activity:activity-ktx:1.12.1'
+    implementation 'androidx.activity:activity-ktx:1.12.2'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.10.0'
     implementation 'androidx.lifecycle:lifecycle-runtime-compose-android:2.10.0'
+    implementation 'androidx.test.ext:junit-ktx:1.3.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:5.0.5'
     implementation 'com.github.chrisbanes:PhotoView:2.3.0'
 
 //  Compose
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.10.0'
-    implementation 'androidx.activity:activity-compose:1.12.1'
-    implementation platform('androidx.compose:compose-bom:2025.12.00')
+    implementation 'androidx.activity:activity-compose:1.12.2'
+    implementation platform('androidx.compose:compose-bom:2025.12.01')
     implementation 'androidx.compose.ui:ui'
     implementation 'androidx.compose.ui:ui-graphics'
     implementation 'androidx.compose.ui:ui-tooling-preview'
@@ -95,7 +98,7 @@ dependencies {
     implementation "androidx.compose.runtime:runtime-livedata:1.10.0"
     implementation 'androidx.compose.material:material-icons-extended-android:1.7.8'
     debugImplementation 'androidx.compose.ui:ui-test-manifest'
-    androidTestImplementation platform('androidx.compose:compose-bom:2025.12.00')
+    androidTestImplementation platform('androidx.compose:compose-bom:2025.12.01')
     androidTestImplementation 'androidx.compose.ui:ui-test-junit4'
 
 //  For Camera
@@ -118,6 +121,11 @@ dependencies {
     implementation "androidx.preference:preference-ktx:1.2.1"
 
     debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.14'
+
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation "androidx.test.ext:junit:1.3.0"
+    androidTestImplementation "androidx.test:runner:1.7.0"
+    androidTestImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.10.2"
 }
 
 // https://f-droid.org/docs/Reproducible_Builds/#bug-baselineprofm-not-deterministic
