+++ app/src/main/res/values/strings.xml
@@ -112,6 +112,7 @@
     <string name="notif_exp_file_caption_s">was successfully copied</string>
     <string name="notif_exp_file_caption_e">Failed to copy</string>
     <string name="open_single">Pin this Item</string>
+    <string name="title_activity_auth2">AuthActivity2</string>
 
 
 </resources>
