+++ app/src/main/res/values/themes.xml
@@ -23,7 +23,7 @@
     <style name="Theme.SafeSpace" parent="Theme.Material3.Light.NoActionBar">
         <item name="android:statusBarColor">@color/md_theme_surface</item>
         <item name="android:windowLightStatusBar">true</item>
-<!--        <item name="android:navigationBarColor">@color/md_theme_surface</item>-->
+        <!--        <item name="android:navigationBarColor">@color/md_theme_surface</item>-->
 
         <item name="colorPrimary">@color/md_theme_primary</item>
         <item name="colorOnPrimary">@color/md_theme_onPrimary</item>
